+++ CHANGELOG.md
@@ -37,3 +37,19 @@
 - Nothing
 
 <!-- ppppppppppppppppppppppppppppppppppppppppppppppppppppppppppppppppppppppppppppppppppppppppppppp -->
+
+## [1.1.0] - 2025-08-14
+
+### Added
+
+- 完成vercel部署的所有准备，包括依赖、api/index.js、vercel.json
+
+### Changed
+
+- index.js稍作修改，导出了app给vercel使用
+
+### Fixed
+
+- Nothing
+
+<!-- ppppppppppppppppppppppppppppppppppppppppppppppppppppppppppppppppppppppppppppppppppppppppppppp -->